% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ls.objects}
\alias{ls.objects}
\title{A Memory Object Browser Function
Shamelessly Copied from this stackoverflow.com post
http://stackoverflow.com/questions/1358003/tricks-to-manage-the-available-memory-in-an-r-session}
\usage{
ls.objects(pos = 1, pattern, order.by, decreasing = FALSE, head = FALSE,
  n = 5)
}
\description{
A Memory Object Browser Function
Shamelessly Copied from this stackoverflow.com post
http://stackoverflow.com/questions/1358003/tricks-to-manage-the-available-memory-in-an-r-session
}
\examples{
lsos()
}
\keyword{memory}

