% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rankTab}
\alias{rankTab}
\title{A Function to rank a table() output from most frequently occuring to least.}
\usage{
rankTab(x)
}
\arguments{
\item{x}{A vector}
}
\description{
A Function to rank a table() output from most frequently occuring to least.
}
\examples{
rankTab(iris$Species)
}
\keyword{table}

